requirements.txt
watchdog
pystray
Pillow
schedule
flask
pyinstaller
